Add files via upload
Removed Minimal dificulty, if your using this tool your not playing on Minimal. 
Added weights to the difficulty so that Substantial and Ruthless are chosen slightly more often.
Fixed a visual bug with the quote section not properly resetting. 

Removed the exe for now. Need to rework it a bit.